Changed signed to unsigend types

+++ CodeFusion.ASM/Parsing/CodeUnit.cs
@@ -8,18 +8,18 @@ public struct CodeUnit
 {
     public readonly SourceFile source;
     public readonly List<Inst> insts;
-    public readonly Dictionary<string, long> labels;
-    public readonly Dictionary<long, Token> unresolved;
-    public readonly Dictionary<Word, int> pool;
-    public readonly long addressOffset;
+    public readonly Dictionary<string, ulong> labels;
+    public readonly Dictionary<ulong, Token> unresolved;
+    public readonly Dictionary<Word, ushort> pool;
+    public readonly ulong addressOffset;
 
-    public CodeUnit(SourceFile source, long addressOffset)
+    public CodeUnit(SourceFile source, ulong addressOffset)
     {
         this.source = source;
         this.insts = new List<Inst>();
-        this.labels = new Dictionary<string, long>();
-        this.unresolved = new Dictionary<long, Token>();
-        this.pool = new Dictionary<Word, int>();
+        this.labels = new Dictionary<string, ulong>();
+        this.unresolved = new Dictionary<ulong, Token>();
+        this.pool = new Dictionary<Word, ushort>();
         this.addressOffset = addressOffset;
     }
 }
